updated department service

+++ department-api/src/main/java/com/department/repository/DepartmentRepository.java
@@ -1,5 +1,6 @@
 package com.department.repository;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
@@ -9,7 +10,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.department.entity.Department;
 
 @Repository
-public interface DepartmentRepository extends MongoRepository<Department, Long>{
+public interface DepartmentRepository extends MongoRepository<Department, BigInteger>{
 
 	@Transactional(readOnly = true)
 	// @Query("SELECT s FROM Department s where s.name = ?1")
